remove api and github PAT logic
